# docs-site/openapi/public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feedling API",
    "version": "v1",
    "description": "HTTP API for Feedling accounts, encrypted chat, agent memory, model routing, perception, and proactive experiences."
  },
  "servers": [
    {
      "url": "https://api.feedling.app",
      "description": "Production"
    },
    {
      "url": "https://test-api.feedling.app",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "RuntimeTokenAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/chat/history": {
      "get": {
        "summary": "Chat History",
        "operationId": "chat_history_v1_chat_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum messages in this page.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 200
            },
            "example": 50
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "description": "Lossless forward cursor. Return messages whose durable append sequence is greater than this value; takes precedence over timestamp cursors.",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "example": 1042
          },
          {
            "name": "before_seq",
            "in": "query",
            "required": false,
            "description": "Lossless backward cursor. Return messages whose durable append sequence is less than this value; takes precedence over timestamp cursors.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1042
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Return messages with ts strictly greater than this watermark.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1783962000.0
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return older messages with ts strictly less than this watermark; takes precedence over since.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1783962000.0
          },
          {
            "name": "include_image_body",
            "in": "query",
            "required": false,
            "description": "Set false to omit image and oversized inline bodies.",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": false
          },
          {
            "name": "include_image_bodies",
            "in": "query",
            "required": false,
            "description": "Compatibility alias for include_image_body.",
            "schema": {
              "type": "boolean"
            },
            "deprecated": true
          }
        ],
        "x-feedling-contract-level": "documented",
        "description": "Read encrypted chat history. Use oldest_seq as before_seq for lossless older paging and latest_seq as after_seq for lossless forward paging; timestamp watermarks remain for compatibility."
      },
      "delete": {
        "summary": "Chat History Clear",
        "operationId": "chat_history_clear_v1_chat_history_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatHistoryClearRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/chat/message": {
      "post": {
        "summary": "Chat Message",
        "operationId": "chat_message_v1_chat_message_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTransportRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Store a user chat message as a v1 ciphertext envelope; the server never decrypts it. If the envelope carries a content_pk_fpr label that does not match the user's currently registered content key, the write is rejected with 409 content_pk_fpr_mismatch (re-fetch whoami and re-seal); unlabeled envelopes are accepted for compatibility."
      }
    },
    "/v1/chat/messages/{message_id}/body": {
      "get": {
        "summary": "Chat Message Body",
        "operationId": "chat_message_body_v1_chat_messages__message_id__body_get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or request parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "x-feedling-contract-level": "documented"
      }
    },
...
    "/v1/memory/list": {
      "get": {
        "summary": "Memory List",
        "operationId": "memory_list_v1_memory_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Recommended page size; values above 200 are capped.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "example": 50
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Return memories whose occurred_at is at or after this ISO-8601 value.",
            "schema": {
              "type": "string"
            },
            "example": "2026-07-01T00:00:00Z"
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "description": "Include archived memories.",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "example": false
          }
        ],
        "x-feedling-contract-level": "documented"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericJsonResponse": {
        "type": "object",
        "description": "JSON response object. Endpoint-specific fields may be added compatibly.",
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "description": "Machine-readable error slug, or the legacy structured MCP error.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "required": [
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "type": "string"
                  },
                  "detail": {
                    "type": "string"
                  }
                },
                "additionalProperties": true
              }
            ]
          },
          "detail": {
            "description": "Human-readable context or structured validation details; do not branch on prose.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "blame": {
            "type": "string",
            "enum": [
              "user_provider",
              "provider_transient",
              "user_environment",
              "system"
            ],
            "description": "Stable responsibility classification when the endpoint can identify it."
          },
          "request_id": {
            "type": "string",
            "description": "Support correlation identifier when available."
          }
        },
        "additionalProperties": true,
        "example": {
          "error": "invalid_payload",
          "detail": "The request body is invalid",
          "request_id": "req_abc123"
        }
      },
...
      "ChatTransportRequest": {
        "type": "object",
        "required": [
          "envelope"
        ],
        "properties": {
          "envelope": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          },
          "client_msg_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional logical-send identifier. Reusing it for the same user within 600 seconds returns the first stored row and emits no second consumer wake."
          },
          "content_type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file"
            ],
            "default": "text"
          },
          "file_name": {
            "type": "string"
          },
          "file_mime": {
            "type": "string"
          },
          "caption_envelope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EncryptedEnvelope"
              }
            ],
            "description": "Optional separately-encrypted user text sent alongside an image/file (content_type image or file). Same E2E envelope shape as the main `envelope`; the enclave decrypts it into the message's plaintext content so the agent sees the caption. Ignored for content_type=text."
          }
        },
        "additionalProperties": true
      },
...
      "ChatHistoryClearRequest": {
        "type": "object",
        "required": [
          "confirm"
        ],
        "properties": {
          "confirm": {
            "type": "string",
            "const": "clear-chat-history"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived Feedling user API key."
      },
      "RuntimeTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Feedling-Runtime-Token",
        "description": "Short-lived scoped token used by hosted runtimes."
      },
      "RelayTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relay-Token",
        "description": "Notify-relay token (nrt_…) issued by POST /v1/notify-relay/register. Not an account credential."
      }
    }
  }
}